backend/requirements.txt
aiosmtplib==5.1.0
alembic==1.13.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
cryptography==46.0.7
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
Faker==25.2.0
fastapi==0.111.0
fastapi-cli==0.0.24
fastapi-mail==1.6.2
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
llvmlite==0.47.0
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numba==0.65.0
numpy==1.26.4
openai==1.30.1
orjson==3.11.8
packaging==26.1
pandas==2.2.2
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.52
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
pytest==8.2.0
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
redis==5.0.4
regex==2025.11.3
rich==15.0.0
rich-toolkit==0.19.7
rsa==4.9.1
scikit-learn==1.4.2
scipy==1.17.1
shap==0.45.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.7
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ujson==5.12.0
uvicorn==0.29.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
xgboost==2.0.3
prometheus-fastapi-instrumentator
psycopg2-binary
